+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Added support for handling unknown variants for `account::Status`,
+  `asset::Exchange`, `events::TradeStatus`, and `order::Status`
 - Removed serialization support for `Exchange` enum